Fix back-compat issue in physics (es6)
